refactor: :art: Rename `loaders` to `extract`
